+++ src/components/button.tsx
@@ -20,6 +20,7 @@
 
 import { AiOutlineMail } from 'react-icons/ai'
 import ReactTooltip from 'react-tooltip'
+import CircularProgress from '@mui/material/CircularProgress';
 
 const Button = ({
   label,
@@ -29,44 +30,66 @@ const Button = ({
   showTooltip = false,
   tooltipText = '',
   disabled = false,
+  loading = false,
 }) => {
-  return showTooltip ? (
+  const renderButton = () => {
+    return loading ?
+      <button className={styleClass} onClick={handleClick} disabled={disabled} style={{ display: 'flex', alignItems: 'center' }}>
+        {
+          <>
+            <CircularProgress
+              sx={{ color: '#fff', marginRight: '5px' }}
+              size={20}
+            />
+            {label}
+          </>
+        }
+      </button>
+      :
+      <button className={styleClass} onClick={handleClick} disabled={disabled}>
+        {!icon ? (
+          label
+        ) : (
+          <>
+            <AiOutlineMail className="button-icon" /> <span>{label}</span>
+          </>
+        )}
+      </button>
+  }
+  return (
     <>
-      <ReactTooltip
-        className="tooltip"
-        id="tooltipBtn"
-        place="top"
-        effect="solid"
-      >
-        {tooltipText}
-      </ReactTooltip>
-      <div data-tip data-for="tooltipBtn">
-        <button
-          className={styleClass}
-          onClick={handleClick}
-          color="#CBCBCB"
-          disabled
-        >
-          {!icon ? (
-            label
-          ) : (
-            <>
-              <AiOutlineMail className="button-icon" /> <span>{label}</span>
-            </>
-          )}
-        </button>
-      </div>
+      {
+        showTooltip ?
+          <>
+            <ReactTooltip
+              className="tooltip"
+              id="tooltipBtn"
+              place="top"
+              effect="solid"
+            >
+              {tooltipText}
+            </ReactTooltip>
+            <div data-tip data-for="tooltipBtn">
+              <button
+                className={styleClass}
+                onClick={handleClick}
+                color="#CBCBCB"
+                disabled
+              >
+                {!icon ? (
+                  label
+                ) : (
+                  <>
+                    <AiOutlineMail className="button-icon" /> <span>{label}</span>
+                  </>
+                )}
+              </button>
+            </div>
+          </>
+          :
+          renderButton()
+      }
     </>
-  ) : (
-    <button className={styleClass} onClick={handleClick} disabled={disabled}>
-      {!icon ? (
-        label
-      ) : (
-        <>
-          <AiOutlineMail className="button-icon" /> <span>{label}</span>
-        </>
-      )}
-    </button>
   )
 }